Revert users.id to INT
This reverts commit aa26d648c447075ac6cbf0dd3bd1ea5051246725.

+++ service/src/main/java/com/codeforcommunity/processor/ProtectedUserProcessorImpl.java
@@ -30,7 +30,7 @@ public class ProtectedUserProcessorImpl implements IProtectedUserProcessor {
 
   @Override
   public void deleteUser(JWTData userData) {
-    long userId = userData.getUserId();
+    int userId = userData.getUserId();
 
     db.deleteFrom(VERIFICATION_KEYS).where(VERIFICATION_KEYS.USER_ID.eq(userId)).executeAsync();
